+++ media/views/__init__.py
@@ -1,0 +1,2 @@
+from .bucket import create_bucket, get_bucket, list_buckets
+from .object import list_objects, upload_file, delete_objects
