+++ snaplapse/settings.py
@@ -42,8 +42,7 @@ INSTALLED_APPS = [
     'django.contrib.contenttypes',
     'django.contrib.sessions',
     'django.contrib.messages',
-    'django.contrib.staticfiles',
-    'storages',
+    'django.contrib.staticfiles'
 ]
 
 MIDDLEWARE = [
